+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/execution/debugger/RobotDebugRunner.java
@@ -2,25 +2,28 @@ package dev.xeonkryptos.xeonrobotframeworkplugin.execution.debugger;
 
 import com.intellij.execution.ExecutionException;
 import com.intellij.execution.ExecutionResult;
+import com.intellij.execution.Executor;
 import com.intellij.execution.configurations.RunProfile;
 import com.intellij.execution.configurations.RunProfileState;
 import com.intellij.execution.configurations.RunnerSettings;
 import com.intellij.execution.executors.DefaultDebugExecutor;
 import com.intellij.execution.runners.ExecutionEnvironment;
+import com.intellij.execution.runners.ExecutionEnvironmentBuilder;
 import com.intellij.execution.runners.ProgramRunner;
-import com.intellij.execution.ui.RunContentDescriptor;
 import com.intellij.xdebugger.XDebugSession;
 import com.jetbrains.python.debugger.PyDebugProcess;
 import com.jetbrains.python.debugger.PyDebugRunner;
+import com.jetbrains.python.parser.icons.PythonParserIcons;
 import com.jetbrains.python.run.PythonCommandLineState;
-import dev.xeonkryptos.xeonrobotframeworkplugin.execution.RobotCommandLineState;
 import dev.xeonkryptos.xeonrobotframeworkplugin.execution.RobotPythonCommandLineState;
 import dev.xeonkryptos.xeonrobotframeworkplugin.execution.config.RobotRunConfiguration;
+import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-import org.jetbrains.concurrency.Promise;
 
+import javax.swing.Icon;
+import java.lang.reflect.Method;
 import java.net.ServerSocket;
 
 public class RobotDebugRunner implements ProgramRunner<RunnerSettings> {
@@ -39,51 +42,61 @@ public class RobotDebugRunner implements ProgramRunner<RunnerSettings> {
 
     @Override
     public void execute(@NotNull ExecutionEnvironment environment) throws ExecutionException {
-        new CustomPyDebugRunner().execute(environment);
+        RobotRunConfiguration robotRunConfiguration = (RobotRunConfiguration) environment.getRunProfile();
+        RunProfile pythonRunProfile = new CustomPythonRunProfile(robotRunConfiguration);
+        ExecutionEnvironment pythonEnvironment = new ExecutionEnvironmentBuilder(environment).runProfile(pythonRunProfile).build();
+        new CustomPyDebugRunner().execute(pythonEnvironment);
+    }
+
+    @RequiredArgsConstructor
+    private static class CustomPythonRunProfile implements RunProfile {
+
+        private final RobotRunConfiguration robotRunConfiguration;
+
+        private RobotPythonCommandLineState state;
+
+        @Override
+        public RunProfileState getState(@NotNull Executor executor, @NotNull ExecutionEnvironment environment) {
+            if (state == null) {
+                state = new RobotPythonCommandLineState(robotRunConfiguration, environment);
+            }
+            return state;
+        }
+
+        @Override
+        public @NotNull String getName() {
+            return "";
+        }
+
+        @Override
+        public Icon getIcon() {
+            return PythonParserIcons.PythonFile;
+        }
     }
 
     private static class CustomPyDebugRunner extends PyDebugRunner {
 
-        private RobotPythonCommandLineState robotPythonCommandLineState;
-
         @NotNull
         @Override
-        protected Promise<@Nullable RunContentDescriptor> execute(@NotNull ExecutionEnvironment environment, @NotNull RunProfileState state) throws
-                                                                                                                                             ExecutionException {
-            if (state instanceof RobotCommandLineState robotState) {
-                state = new RobotPythonCommandLineState(robotState.getRobotRunConfiguration(), environment);
-            }
-            robotPythonCommandLineState = (RobotPythonCommandLineState) state;
-            return super.execute(environment, state);
-        }
-
-        @Override
-        protected Promise<@NotNull XDebugSession> createSession(@NotNull RunProfileState state, @NotNull ExecutionEnvironment environment) {
-            if (state instanceof RobotCommandLineState robotState) {
-                state = new RobotPythonCommandLineState(robotState.getRobotRunConfiguration(), environment);
-            }
-            robotPythonCommandLineState = (RobotPythonCommandLineState) state;
-            return super.createSession(state, environment);
+        protected PyDebugProcess createDebugProcess(@NotNull XDebugSession session, ServerSocket serverSocket, ExecutionResult result, PythonCommandLineState pyState) {
+            return new RobotPyDebugProcess(session, serverSocket, result.getExecutionConsole(), result.getProcessHandler(), pyState.isMultiprocessDebug(), (RobotPythonCommandLineState) pyState);
         }
 
         @NotNull
         @Override
-        protected PyDebugProcess createDebugProcess(@NotNull XDebugSession session,
-                                                    ServerSocket serverSocket,
-                                                    ExecutionResult result,
-                                                    PythonCommandLineState pyState) {
-            return new RobotPyDebugProcess(session,
-                                           serverSocket,
-                                           result.getExecutionConsole(),
-                                           result.getProcessHandler(),
-                                           pyState.isMultiprocessDebug(),
-                                           robotPythonCommandLineState);
-        }
-
-        @NotNull
-        @Override
+        @SneakyThrows
         protected PyDebugProcess createDebugProcess(@NotNull XDebugSession session, int serverPort, ExecutionResult result) {
-            return new RobotPyDebugProcess(session, result, serverPort, robotPythonCommandLineState);
+            RunProfile runProfile = session.getRunProfile();
+            assert runProfile != null;
+
+            // Have to resolve to reflection here as the method itself is public, but within the implementing class, not in the interface itself. In future versions it changes, too, but for now
+            // this is a necessary workaround until upgrading to a higher platform version
+            Method getExecutionEnvironment = session.getClass().getMethod("getExecutionEnvironment");
+            ExecutionEnvironment executionEnvironment = (ExecutionEnvironment) getExecutionEnvironment.invoke(session);
+
+            RobotPythonCommandLineState state = (RobotPythonCommandLineState) runProfile.getState(executionEnvironment.getExecutor(), executionEnvironment);
+            assert state != null;
+            return new RobotPyDebugProcess(session, result, serverPort, state);
         }
     }
 }