+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Execution directory doesn't change at runtime from configured or evaluated working directory with pyproject.toml
 - Parsing of conditional statements consisting of a variable only
 - Code completions for Test Setup, Task Setup, Test Teardown, Task Teardown are now available again
+- ClassCastException when starting a robot debug process
 
 ## 0.7.4 - 2026-06-06